+++ HuffmanCode-iOS/API/HuffmanAPI.swift
@@ -1,5 +1,5 @@
 //
-//  API.swift
+//  HuffmanAPI.swift
 //  HuffmanCode-iOS
 //
 //  Created by 宮澤賢生 on 2017/09/09.
@@ -9,44 +9,6 @@
 import Foundation
 import APIKit
 
-protocol HuffmanRequest: Request {
-}
-extension HuffmanRequest {
-    var baseURL: URL {
-        return URL(string: "http://localhost:8000")!
-    }
-}
-
-extension HuffmanRequest where Response: Decodable {
-    var dataParser: DataParser {
-        return DecodableDataParser()
-    }
-    
-    
-    func response(from object: Any, urlResponse: HTTPURLResponse) throws -> Response {
-        guard let data = object as? Data else {
-            throw ResponseError.unexpectedObject(object)
-        }
-        return try JSONDecoder().decode(Response.self, from: data)
-    }
-}
-
-struct HuffmanResponse: Decodable {
-    let message: String
-    private enum CodingKeys: String, CodingKey {
-        case message
-    }
-}
-class DecodableDataParser: DataParser {
-    var contentType: String? {
-        return "application/json"
-    }
-    
-    func parse(data: Data) throws -> Any {
-        return data
-    }
-}
-
 class HuffmanAPI {
     private init() {}
     
@@ -55,21 +17,10 @@ class HuffmanAPI {
         let method: HTTPMethod = .post
         let path: String = "/huffman"
         var parameters: Any? {
-            return ["data": Huffman3.share.freq]
+            return ["data": HuffmanModel.share.frequency]
         }
     }
 }
 
-class HuffmanSession<T: Request>{
-    func session(request: T) {
-        Session.send(request) { result in
-            switch result {
-            case .success(let response):
-                print(response)
-            case .failure(let error):
-                print(error)
-            }
-        }
-    }
-}
+
