+++ HuffmanCode-iOS/API/Model/HuffmanResponse.swift
@@ -7,3 +7,11 @@
 //
 
 import Foundation
+import APIKit
+
+struct HuffmanResponse: Decodable {
+    let message: String
+    private enum CodingKeys: String, CodingKey {
+        case message
+    }
+}
